memory/requirements.txt
langchain
langchain-google-genai
faiss-cpu 
google-generativeai
python-dotenv
langchain-community
sentence-transformers
torch